refactor(platform): split index.ts god-file into cohesive sibling modules
Export-surface-stable split of packages/platform/src/index.ts (~1111 lines,
one file, ~5 unrelated concerns) into cohesive sibling modules, with index.ts
kept as a pure barrel of explicit named re-exports.

- command.ts    cross-platform command-invocation construction
- process.ts    process lifecycle, stamps, snapshots, stop escalation
- proxy-env.ts  system proxy discovery + proxy-aware env merging
- fs.ts         path containment, atomic copy, removal, log tails
- http.ts       HTTP readiness polling
- toolchain.ts  user-level toolchain bin discovery

Function bodies moved byte-identically; only import paths adjusted. Added a
@module docblock per file and JSDoc to every exported function. Tiny private
error-extraction helpers (errorCode/errorMessage) are duplicated where used to
keep the public surface identical (no new exported names). Public export set
unchanged (42 names); typecheck + 76 tests + build all green.

Tracked by #5165.
